Feature tests (#32)
* add config test

* fix tests

* fix migrations

* fix create default value

* simplify code

* small extend test-case

* small fixes

+++ database/migrations/2016_06_24_094039_create_configs_table.php
@@ -16,7 +16,7 @@ class CreateConfigsTable extends Migration
         DB::getDoctrineSchemaManager()->getDatabasePlatform()->registerDoctrineTypeMapping('jsonb', 'array');
         Schema::create('configs', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('chat_id')->unique();
+            $table->unsignedBigInteger('chat_id')->unique();
             $table->jsonb('config');
             $table->timestamps();
         });
